BRD: Weaving Logic/Checks
- Added A "System" That Checks The Last Two Spells Casted And Determines If Those Were oGCD Spells

SideNote : Even tho this is somewhat "complicated", this is in my opinion the most flexible system. It wont break with custom openers and it allows us to tripple weave some cases if we rly want to. Checking for GCD TimeLeft wouldn't let us do that and had to be fine tuned for everyone on their end. This solution should work out of the box and for every ping.