Make Magitek Great Again
Fixed most of SMN's issues with the help of Rhokzor. If there's anything missed, let Sage/Omni know.

Fixed DNC wasting a feather in aoes

Added some checks to various routines to not waste CDs on invulnerable enemies. Might do a more extensive, smater version of this if it's actually useful.