Minor fixes to redraw and divination
Previously seemed to not try for seals when redraw was running low on charges, changed to check if greater than OR equal to 1 charge.

Added setting for divination allies so it doesnt go off unless configured number (default 2) allies are in range.

To do:
Add configurable value to the GUI