Dancer - Manger Dance Change
This reworks the dance system. Prior dancing would break at times and get lost on what step it was doing. Normally if you left and reentered combat mid dance.

This has stops stops the routine from completely the dance when not in target range. Some boss fights force you out of range.

This should now put it in state to test for raids.