Address a decision issue in Autonomous mode
This should help with the issue where the bot keeps running after doing an avoid. 

Here is the logic sequence:

* Combat Handed off to Magitek
* Avoidance manager takes over (triggered by SideStep or w/e) Interupting an ability
* Avoidance concludes, but doesn't stop movement (we didn't make it out of the avoid, or in some cases, did make it out of the avoid but didn't stop) 
* Magitek (particularly casters) were stuck in a loop trying to cast but failing due to `IsMoving`.
* bot runs into wall (or w/e)

This change:

This change basically adjusts the return conditional to be true when autonomous. (Eg we don't care if we have swiftcast or are moving) and in autonomous mode we don't care about the adjust cast time. (TBH I think that can be removed).


I have tested against AST, and I haven't seen the issue re-appear. Going to be testing on CNJ again in a bit. 


---

TODO: 

Improvements still need to be made (especially to healers) for LOS checking of targets. I am not sure why, `CanCastOrQueue` is returning true, but going to cast the ability returns fails in some odd edge cases.